Add ValidatorConfig
